Vote logic now works correctly

+++ website/DBmanager.py
@@ -54,7 +54,7 @@ class DBManager:
         upvotes=self.interface.cur.fetchone()
         self.interface.execute("SELECT COUNT(*) FROM user_votes WHERE question_id=%s AND vote=0", (question_ID))
         downvotes=self.interface.cur.fetchone()
-        return upvotes-downvotes
+        return upvotes[0]-downvotes[0]
     
     def hasVoted(self, question_ID, user_ID):
         self.interface.execute("SELECT vote FROM user_votes WHERE question_ID=%s AND user_id=%s", (question_ID, user_ID))
@@ -65,15 +65,29 @@ class DBManager:
             return True
     
     def submitVote(self, question_id, user_ID, vote):
+        if vote not in [0,1]:
+            self.interface.execute("SELECT rating FROM user_questions WHERE question_id=%s", (question_id))
+            rating=self.interface.cur.fetchone()
+            return rating
         if self.hasVoted(question_id, user_ID)==False:
             self.interface.execute("INSERT INTO user_votes (question_id, user_id,vote) VALUES (%s,%s,%s)", (question_id,user_ID,vote))
         else:
             self.interface.execute("SELECT vote FROM user_votes WHERE question_ID=%s AND user_id=%s", (question_id, user_ID))
             row=self.interface.cur.fetchone()
-            if vote==0:                                                                                                                    #Updating rating to reflect new state
+            if row[0]==0:                                                                                                                    #Updating rating to reflect new state
                 self.interface.execute("UPDATE user_questions SET rating = rating + 1 WHERE question_id=%s", (question_id))
-            elif vote==1:
+                if vote==0:
+                    self.interface.execute("DELETE FROM user_votes WHERE question_id=%s AND user_id=%s", (question_id,user_ID)) 
+                    self.interface.execute("SELECT rating FROM user_questions WHERE question_id=%s", (question_id))
+                    rating=self.interface.cur.fetchone()
+                    return rating
+            elif row[0]==1:
                 self.interface.execute("UPDATE user_questions SET rating = rating - 1 WHERE question_id=%s", (question_id))
+                if vote==1:
+                    self.interface.execute("DELETE FROM user_votes WHERE question_id=%s AND user_id=%s", (question_id,user_ID)) 
+                    self.interface.execute("SELECT rating FROM user_questions WHERE question_id=%s", (question_id))
+                    rating=self.interface.cur.fetchone()
+                    return rating
 
             self.interface.execute("DELETE FROM user_votes WHERE question_id=%s AND user_id=%s", (question_id,user_ID))                     #Clearing out the old vote
             self.interface.execute("INSERT INTO user_votes (question_id, user_id,vote) VALUES (%s,%s,%s)", (question_id,user_ID,vote))      #New vote
@@ -81,6 +95,9 @@ class DBManager:
             self.interface.execute("UPDATE user_questions SET rating = rating - 1 WHERE question_id=%s", (question_id))                     #Making rating reflect new vote
         elif vote==1:
             self.interface.execute("UPDATE user_questions SET rating = rating + 1 WHERE question_id=%s", (question_id))
+        self.interface.execute("SELECT rating FROM user_questions WHERE question_id=%s", (question_id))
+        rating=self.interface.cur.fetchone()
+        return rating
